[ckcore] Restructure cli package
- init: contains type vars and global functions
- commands: holds all commands (except help, which is special)
- cli: has parser and interpreter
