+++ ckcore/tests/core/cli/command_test.py
@@ -8,10 +8,10 @@ from _pytest.logging import LogCaptureFixture
 from aiostream import stream
 from pytest import fixture
 
-from core.cli.cli import CLI, CLIDependencies
+from core.cli.cli import CLI
 
 # noinspection PyUnresolvedReferences
-from core.cli.command import ListSink
+from core.cli.command import ListSink, CLIDependencies
 from core.db.jobdb import JobDb
 from core.error import CLIParseError
 from core.model.model import predefined_kinds
